From Jordi Torres, "This error occurs when compiling current OSG with the NDK >=r8c:
android-ndk-r8d/build/gmsl/__
gmsl:512: *** non-numeric second argument to `wordlist' function: ''. Stop.

This is discussed in this thread[1].
The solution is to add <uses-sdk android:minSdkVersion="8" /> to the manifest as the OSG library will only run in devices with Android>=2.2 (Froyo) with API level 8 [2].

The fix does not affect to older sdk's.

Thanks to Jan Ciger for spotting the bug and find a possible solution. "


git-svn-id: http://svn.openscenegraph.org/osg/OpenSceneGraph/trunk@13346 16af8721-9629-0410-8352-f15c8da7e697